+++ sora/occultation/core.py
@@ -1,19 +1,20 @@
-from .star import Star
-from .ephem import EphemPlanete, EphemJPL, EphemKernel
-from .observer import Observer
-from .lightcurve import LightCurve
-from .prediction import occ_params, PredictionTable
-from .extra import ChiSquare
+from .chordlist import ChordList
+from sora.star import Star
+from sora.ephem import EphemPlanete, EphemJPL, EphemKernel, EphemHorizons
+from sora.observer import Observer
+from sora.prediction import occ_params, PredictionTable
+from sora.extra import ChiSquare
 from sora.body import Body
-from sora.config.decorators import deprecated_alias
+from sora.config.decorators import deprecated_alias, deprecated_function
 import astropy.units as u
 from astropy.coordinates import SkyCoord, SkyOffsetFrame
 from astropy.time import Time
 import numpy as np
 import warnings
-import matplotlib.pyplot as plt
+from functools import partial
 
 
+__all__ = ['fit_ellipse', 'Occultation', 'positionv']
 warnings.simplefilter('always', UserWarning)
 
 
@@ -34,7 +35,7 @@ def positionv(star, ephem, observer, time):
     """
     if type(star) != Star:
         raise ValueError('star must be a Star object')
-    if type(ephem) not in [EphemPlanete, EphemJPL, EphemKernel]:
+    if type(ephem) not in [EphemPlanete, EphemJPL, EphemKernel, EphemHorizons]:
         raise ValueError('ephem must be an Ephemeris object')
     if type(observer) != Observer:
         raise ValueError('observer must be an Observer object')
@@ -103,30 +104,23 @@ def fit_ellipse(*args, equatorial_radius, dequatorial_radius=0, center_f=0, dcen
     """
     values = []
     chord_name = []
+    if len(args) == 0:
+        raise ValueError('No occultation have been given as input.')
     for occ in args:
-        if type(occ) != Occultation:
+        if not isinstance(occ, Occultation):
             raise TypeError('Given argument must be an Occultation object.')
-        pos = occ.positions
-        for site in pos.keys():
-            pos_obs = pos[site]
-            for lc in pos_obs.keys():
-                pos_lc = pos_obs[lc]
-                if type(pos_lc) != _PositionDict:
-                    continue
-                if pos_lc['status'] == 'positive':
-                    if pos_lc['immersion']['on']:
-                        f, g = pos_lc['immersion']['value']
-                        err = np.array(pos_lc['immersion']['error'])
-                        erro = np.linalg.norm(err[0]-err[1])/2.0
-                        values.append([f, g, erro])
-                        chord_name.append(lc.replace(' ', '_') + '_immersion')
-
-                    if pos_lc['emersion']['on']:
-                        f, g = pos_lc['emersion']['value']
-                        err = np.array(pos_lc['emersion']['error'])
-                        erro = np.linalg.norm(err[0]-err[1])/2.0
-                        values.append([f, g, erro])
-                        chord_name.append(lc.replace(' ', '_') + '_emersion')
+        for name, chord in occ.chords.items():
+            if chord.status() == 'positive':
+                if chord.is_able['immersion']:
+                    f, g, vf, vg = chord.get_fg(time='immersion', vel=True)
+                    err = np.linalg.norm([vf, vg])*chord.lightcurve.immersion_err
+                    values.append([f, g, err])
+                    chord_name.append(name + '_immersion')
+                if chord.is_able['emersion']:
+                    f, g, vf, vg = chord.get_fg(time='emersion', vel=True)
+                    err = np.linalg.norm([vf, vg])*chord.lightcurve.emersion_err
+                    values.append([f, g, err])
+                    chord_name.append(name + '_emersion')
 
     controle_f0 = Time.now()
     f0_chi = np.array([])
@@ -239,6 +233,10 @@ class _PositionDict(dict):
                     self[key][k] = value
             elif key == 'on':
                 super().__setitem__('on', status[value])
+                if status[value]:
+                    self['enable']()
+                else:
+                    self['disable']()
         else:
             if value not in status.keys():
                 raise ValueError("Value must be 'on' or 'off' only.")
@@ -251,8 +249,20 @@ class _PositionDict(dict):
             raise KeyError('Key "{}" does not exist'.format(key))
 
     def __str__(self):
-        out = '\n' + '\n'.join(['{}: {}'.format(key, self[key]) for key in self.keys()])
+        out = []
+        for key in self.keys():
+            if key not in ['enable', 'disable']:
+                out.append('{}: {}'.format(key, self[key]))
+        out = '\n' + '\n'.join(out)
         return out.replace('\n', '\n  ')
+
+    def __repr__(self):
+        out = []
+        for key in self.keys():
+            if key not in ['enable', 'disable']:
+                out.append('\'{}\': {}'.format(key, self[key].__repr__()))
+        out = ',\n'.join(out)
+        return '{' + out.replace('\n', '\n  ') + '}'
 
 
 class Occultation():
@@ -287,19 +297,19 @@ class Occultation():
             star = Star(coord=star)
         elif not isinstance(star, Star):
             raise ValueError('"star" must be a Star object or a string with coordinates of the star')
-        self.star = star
+        self._star = star
         if body is not None:
             if not isinstance(body, (str, Body)):
                 raise ValueError('"body" must be a string with the name of the object or a Body object')
             if isinstance(body, str):
                 body = Body(name=body, mode='sbdb')
-            self.body = body
+            self._body = body
         if ephem is not None:
             if body is not None:
                 self.body.ephem = ephem
             else:
                 if hasattr(ephem, 'name'):
-                    self.body = Body(name=ephem.name, mode='sbdb', ephem=ephem)
+                    self._body = Body(name=ephem.name, mode='sbdb', ephem=ephem)
                 else:
                     raise ValueError('When only "ephem" is given, "ephem" must have a name for search.')
         try:
@@ -325,7 +335,23 @@ class Occultation():
 
         self.__observations = []
         self._position = _PositionDict()
+        self._chords = ChordList(star=self.star, body=self._body, time=self.tca)
+        self._chords._shared_with['occultation'] = {"vel": np.absolute(self.vel), "dist": float(self.dist.AU),
+                                                    "star_diam": float(self.star_diam.km)}
 
+    @property
+    def star(self):
+        return self._star
+
+    @property
+    def body(self):
+        return self._body
+
+    @property
+    def chords(self):
+        return self._chords
+
+    @deprecated_function(message="Please use chords.add_chord to add new observations.")
     def add_observation(self, obs, lightcurve):
         """ Adds observations to the Occultation object.
 
@@ -333,23 +359,9 @@ class Occultation():
             obs (Observer): The Observer object to be added.
             lightcurve (LightCurve): The LightCurve object to be added
         """
-        if type(obs) != Observer:
-            raise ValueError('obs must be an Observer object')
-        if type(lightcurve) != LightCurve:
-            raise ValueError('lightcurve must be a LightCurve object')
-        for o, l in self.__observations:
-            if l.name == lightcurve.name:
-                raise ValueError('{} LightCurve already associated to {} Observer'.format(lightcurve.name, o.name))
-        self.__observations.append((obs, lightcurve))
-        lightcurve.set_vel(np.absolute(self.vel))
-        lightcurve.set_dist(float(self.dist.AU))
-        lightcurve.set_star_diam(float(self.star_diam.km))
-        try:
-            lightcurve.calc_magnitude_drop(mag_star=self.star.mag['G'], mag_obj=self.body.apparent_magnitude(self.tca))
-        except:
-            lightcurve.bottom_flux = 0.0
-            warnings.warn('Magnitude drop was not calculated. Using bottom flux as 0.0 instead.')
+        self.chords.add_chord(name=lightcurve.name, observer=obs, lightcurve=lightcurve)
 
+    @deprecated_function(message="Please use chords.remove_chord to remove observations.")
     def remove_observation(self, key, key_lc=None):
         """ Removes an observation from the Occultation object.
 
@@ -358,43 +370,28 @@ class Occultation():
             keylc (str): In the case where repeated names are present for different observations,
                 keylc must be given for the name of the LightCurve and key will be used for the name of the Observer.
         """
-        rm_list = np.array([])
-        same_key = False
-        if key_lc is None:
-            same_key = True
-            key_lc = key
-        ko = []
-        kl = []
-        for i, val in enumerate(self.__observations):
-            if val[0].name == key:
-                ko.append(i)
-            if val[1].name == key_lc:
-                kl.append(i)
-        ko = np.array(ko)
-        kl = np.array(kl)
-        if not same_key:
-            k = ko[np.where(ko == kl)[0]]
-            rm_list = np.hstack((rm_list, k))
-        else:
-            rm_list = np.hstack((rm_list, np.array(kl)))
-            if len(kl) > 0 and len(ko) > 0 and kl[0] not in ko:
-                raise ValueError("Observation could not univocally be identified, "
-                                 "please give parameters for Observer and LightCurve")
-            rm_list = np.hstack((rm_list, np.array(ko)))
-        rm_list = np.unique(rm_list)
-        if len(rm_list) == 0:
-            raise ValueError('No observer "{}" and/or lightcurve "{}" was found'.format(key, key_lc))
-        list = np.arange(len(self.__observations)).tolist()
-        for i in rm_list:
-            list.remove(i)
-        self.__observations = [self.__observations[item] for item in list]
+        try:
+            self.chords.remove_chord(name=key)
+        except KeyError:
+            rm_list = []
+            for name, chord in self.chords.items():
+                if chord.observer.name == key and (key_lc is None or chord.lightcurve.name == key_lc):
+                    rm_list.append(name)
+            if len(rm_list) != 1:
+                if len(rm_list) == 0:
+                    err_mes = "No observation was identified with given keys."
+                else:
+                    err_mes = "More than one chord was identified. Please provide unique keys."
+                raise ValueError(err_mes)
+            else:
+                self.chords.remove_chord(name=rm_list[0])
 
+    @deprecated_function(message="Please use chords.")
     def observations(self):
         """ Print all the observations added to the Occultation object
             Pair (Observer, LightCurve)
         """
-        for o, l in self.__observations:
-            print('Observer= {}, LC: {}'.format(o.name, l.name))
+        print(self.chords.__repr__())
 
     def fit_ellipse(self, **kwargs):
         """ Fits an ellipse to the chords of this occultation
@@ -431,103 +428,114 @@ class Occultation():
         return chisquare
 
     @property
+    @deprecated_function(message="Please use chords.summary()")
     def positions(self):
         """ Calculates the position and velocity for all chords.
             Saves it into an _PositionDict object.
         """
         position = self._position
-        if len(self.__observations) == 0:
+        if len(self.chords) == 0:
             raise ValueError('There is no observation defined for this occultation')
 
         pair = []
-        for o, l in self.__observations:
-            pair.append((o.name, l.name))
+        for name, chord in self.chords.items():
+            obs = chord.observer
+            obs_name = obs.name
+            if obs_name == '':
+                obs_name = '{}_obs'.format(chord.name)
+            lc = chord.lightcurve
+            lc_name = lc.name
+            if lc_name == '':
+                lc_name = '{}_lc'.format(chord.name)
+            pair.append((obs_name, lc_name))
 
-            coord = [o.lon, o.lat, o.height]
-            if o.name not in position.keys():
-                position['_occ_'+o.name] = _PositionDict(lon=o.lon, lat=o.lat, height=o.height)
-                position[o.name]['_occ_lon'] = o.lon
-                position[o.name]['_occ_lat'] = o.lat
-                position[o.name]['_occ_height'] = o.height
-            pos_obs = position[o.name]
+            coord = [obs.lon, obs.lat, obs.height]
+            if obs.name not in position.keys():
+                position['_occ_'+obs_name] = _PositionDict(lon=obs.lon, lat=obs.lat, height=obs.height)
+                position[obs_name]['_occ_lon'] = obs.lon
+                position[obs_name]['_occ_lat'] = obs.lat
+                position[obs_name]['_occ_height'] = obs.height
+            pos_obs = position[obs_name]
             coord2 = [pos_obs['lon'], pos_obs['lat'], pos_obs['height']]
-            if o.lon != pos_obs['lon']:
-                position[o.name]['_occ_lon'] = o.lon
-            if o.lat != pos_obs['lat']:
-                position[o.name]['_occ_lat'] = o.lat
-            if o.height != pos_obs['height']:
-                position[o.name]['_occ_height'] = o.height
+            if obs.lon != pos_obs['lon']:
+                position[obs_name]['_occ_lon'] = obs.lon
+            if obs.lat != pos_obs['lat']:
+                position[obs_name]['_occ_lat'] = obs.lat
+            if obs.height != pos_obs['height']:
+                position[obs_name]['_occ_height'] = obs.height
             samecoord = (coord == coord2)
 
-            if l.name not in pos_obs.keys():
-                pos_obs['_occ_'+l.name] = _PositionDict()
-            pos_lc = pos_obs[l.name]
+            if lc_name not in pos_obs.keys():
+                pos_obs['_occ_'+lc_name] = _PositionDict()
+            pos_lc = pos_obs[lc_name]
 
-            pos_lc['_occ_status'] = 'negative'
-            if hasattr(l, 'immersion') or hasattr(l, 'emersion'):
-                pos_lc['_occ_status'] = 'positive'
+            pos_lc['_occ_status'] = chord.status()
 
-            if hasattr(l, 'immersion'):
+            if hasattr(lc, 'immersion'):
                 if 'immersion' not in pos_lc.keys():
-                    pos_lc['_occ_immersion'] = _PositionDict(on=True)
+                    pos_lc['_occ_immersion'] = _PositionDict(on=chord.is_able['immersion'],
+                                                             enable=partial(chord.enable, time='immersion'),
+                                                             disable=partial(chord.disable, time='immersion'))
                 obs_im = pos_lc['immersion']
+                obs_im['_occ_on'] = chord.is_able['immersion']
                 do_err = False
-                if samecoord and 'time' in obs_im.keys() and obs_im['time'] == l.immersion:
+                if samecoord and 'time' in obs_im.keys() and obs_im['time'] == lc.immersion:
                     pass
                 else:
                     do_err = True
-                    f1, g1, vf1, vg1 = positionv(self.star, self.body.ephem, o, l.immersion)
-                    obs_im['_occ_time'] = l.immersion
+                    f1, g1, vf1, vg1 = chord.get_fg(time='immersion', vel=True)
+                    obs_im['_occ_time'] = lc.immersion
                     obs_im['_occ_value'] = (round(f1, 3), round(g1, 3))
                     obs_im['_occ_vel'] = (round(vf1, 3), round(vg1, 3))
-                if not do_err and 'time_err' in obs_im.keys() and obs_im['time_err'] == l.immersion_err:
+                if not do_err and 'time_err' in obs_im.keys() and obs_im['time_err'] == lc.immersion_err:
                     pass
                 else:
-                    fe1, ge1 = positionv(self.star, self.body.ephem, o, l.immersion-l.immersion_err*u.s)[0:2]
-                    fe2, ge2 = positionv(self.star, self.body.ephem, o, l.immersion+l.immersion_err*u.s)[0:2]
-                    obs_im['_occ_time_err'] = l.immersion_err
+                    fe1, ge1 = chord.get_fg(time=lc.immersion-lc.immersion_err*u.s)
+                    fe2, ge2 = chord.get_fg(time=lc.immersion+lc.immersion_err*u.s)
+                    obs_im['_occ_time_err'] = lc.immersion_err
                     obs_im['_occ_error'] = ((round(fe1, 3), round(ge1, 3)), (round(fe2, 3), round(ge2, 3)))
 
-            if hasattr(l, 'emersion'):
-                if 'emersion' not in pos_lc.keys():
-                    pos_lc['_occ_emersion'] = _PositionDict(on=True)
+            if hasattr(lc, 'emersion'):
+                pos_lc['_occ_emersion'] = _PositionDict(on=chord.is_able['emersion'],
+                                                        enable=partial(chord.enable, time='emersion'),
+                                                        disable=partial(chord.disable, time='emersion'))
                 obs_em = pos_lc['emersion']
                 do_err = False
-                if samecoord and 'time' in obs_em.keys() and obs_em['time'] == l.emersion:
+                if samecoord and 'time' in obs_em.keys() and obs_em['time'] == lc.emersion:
                     pass
                 else:
                     do_err = True
-                    f1, g1, vf1, vg1 = positionv(self.star, self.body.ephem, o, l.emersion)
-                    obs_em['_occ_time'] = l.emersion
+                    f1, g1, vf1, vg1 = chord.get_fg(time='emersion', vel=True)
+                    obs_em['_occ_time'] = lc.emersion
                     obs_em['_occ_value'] = (round(f1, 3), round(g1, 3))
                     obs_em['_occ_vel'] = (round(vf1, 3), round(vg1, 3))
-                if not do_err and 'time_err' in obs_em.keys() and obs_em['time_err'] == l.emersion_err:
+                if not do_err and 'time_err' in obs_em.keys() and obs_em['time_err'] == lc.emersion_err:
                     pass
                 else:
-                    fe1, ge1 = positionv(self.star, self.body.ephem, o, l.emersion-l.emersion_err*u.s)[0:2]
-                    fe2, ge2 = positionv(self.star, self.body.ephem, o, l.emersion+l.emersion_err*u.s)[0:2]
-                    obs_em['_occ_time_err'] = l.emersion_err
+                    fe1, ge1 = chord.get_fg(time=lc.emersion-lc.emersion_err*u.s)
+                    fe2, ge2 = chord.get_fg(time=lc.emersion+lc.emersion_err*u.s)
+                    obs_em['_occ_time_err'] = lc.emersion_err
                     obs_em['_occ_error'] = ((round(fe1, 3), round(ge1, 3)), (round(fe2, 3), round(ge2, 3)))
 
             if pos_lc['status'] == 'negative':
                 if 'start_obs' not in pos_lc.keys():
-                    pos_lc['_occ_start_obs'] = _PositionDict(on=True)
+                    pos_lc['_occ_start_obs'] = _PositionDict()
                 obs_start = pos_lc['start_obs']
-                if samecoord and 'time' in obs_start.keys() and obs_start['time'] == l.initial_time:
+                if samecoord and 'time' in obs_start.keys() and obs_start['time'] == lc.initial_time:
                     pass
                 else:
-                    f, g, vf, vg = positionv(self.star, self.body.ephem, o, l.initial_time)
-                    obs_start['_occ_time'] = l.initial_time
+                    f, g, vf, vg = chord.get_fg(time='start', vel=True)
+                    obs_start['_occ_time'] = lc.initial_time
                     obs_start['_occ_value'] = (round(f, 3), round(g, 3))
                     obs_start['_occ_vel'] = (round(vf, 3), round(vg, 3))
                 if 'end_obs' not in pos_lc.keys():
-                    pos_lc['_occ_end_obs'] = _PositionDict(on=True)
+                    pos_lc['_occ_end_obs'] = _PositionDict()
                 obs_end = pos_lc['end_obs']
-                if samecoord and 'time' in obs_end.keys() and obs_end['time'] == l.end_time:
+                if samecoord and 'time' in obs_end.keys() and obs_end['time'] == lc.end_time:
                     pass
                 else:
-                    f, g, vf, vg = positionv(self.star, self.body.ephem, o, l.end_time)
-                    obs_end['_occ_time'] = l.end_time
+                    f, g, vf, vg = chord.get_fg(time='end', vel=True)
+                    obs_end['_occ_time'] = lc.end_time
                     obs_end['_occ_value'] = (round(f, 3), round(g, 3))
                     obs_end['_occ_vel'] = (round(vf, 3), round(vg, 3))
 
@@ -563,22 +571,22 @@ class Occultation():
             center = np.array([self.fitted_params['center_f'][0], self.fitted_params['center_g'][0]])
         else:
             center = np.array([0, 0])
-        positions = self.positions
-        for ob, lc in self.__observations:
-            vals = positions[ob.name][lc.name]
-            if all([i not in vals.keys() for i in ['immersion', 'emersion']]):
+        for name, chord in self.chords.items():
+            im = getattr(chord.lightcurve, 'immersion', None)
+            em = getattr(chord.lightcurve, 'emersion', None)
+            if im is None and em is None:
                 continue
-            print('{} - Velocity used: {:.3f}'.format(lc.name, lc.vel))
-            if 'immersion' in vals.keys():
-                im = vals['immersion']
-                delta = np.array(im['value']) - center
+            print('{} - Velocity used: {:.3f}'.format(name, chord.lightcurve.vel))
+            if im is not None:
+                vals = chord.get_fg(time=im, vel=True)
+                delta = np.array(vals[0:2]) - center
                 print('    Immersion Radial Velocity: {:.3f}'.
-                      format(np.abs(np.dot(np.array(im['vel']), delta)/np.linalg.norm(delta))))
-            if 'emersion' in vals.keys():
-                em = vals['emersion']
-                delta = np.array(em['value']) - center
+                      format(np.abs(np.dot(np.array(vals[2:]), delta)/np.linalg.norm(delta))))
+            if em is not None:
+                vals = chord.get_fg(time=em, vel=True)
+                delta = np.array(vals[0:2]) - center
                 print('    Emersion Radial Velocity: {:.3f}'.
-                      format(np.abs(np.dot(np.array(em['vel']), delta)/np.linalg.norm(delta))))
+                      format(np.abs(np.dot(np.array(vals[2:]), delta)/np.linalg.norm(delta))))
 
     def new_astrometric_position(self, time=None, offset=None, error=None, log=True):
         """ Calculates the new astrometric position for the object given fitted parameters
@@ -612,15 +620,12 @@ class Occultation():
         if offset is not None:
             off_ra = offset[0]*u.Unit(offset[2])
             off_dec = offset[1]*u.Unit(offset[2])
-            try:
-                teste = off_ra.to(u.km)
+            if off_ra.unit.physical_type == 'length':
                 dist = True
-            except:
-                try:
-                    teste = off_ra.to(u.arcsec)
-                    dist = False
-                except:
-                    raise ValueError('Offset unit must be a distance or angular value.')
+            elif off_ra.unit.physical_type == 'angle':
+                dist = False
+            else:
+                raise ValueError('Offset unit must be a distance or angular value.')
         elif hasattr(self, 'fitted_params'):
             off_ra = self.fitted_params['center_f'][0]*u.km
             off_dec = self.fitted_params['center_g'][0]*u.km
@@ -634,15 +639,12 @@ class Occultation():
         if error is not None:
             e_off_ra = error[0]*u.Unit(error[2])
             e_off_dec = error[1]*u.Unit(error[2])
-            try:
-                teste = e_off_ra.to(u.km)
+            if e_off_ra.unit.physical_type == 'length':
                 e_dist = True
-            except:
-                try:
-                    teste = e_off_ra.to(u.arcsec)
-                    e_dist = False
-                except:
-                    raise ValueError('Error unit must be a distance or angular value.')
+            elif e_off_ra.unit.physical_type == 'angle':
+                e_dist = False
+            else:
+                raise ValueError('Error unit must be a distance or angular value.')
         elif hasattr(self, 'fitted_params'):
             e_off_ra = self.fitted_params['center_f'][1]*u.km
             e_off_dec = self.fitted_params['center_g'][1]*u.km
@@ -684,6 +686,7 @@ class Occultation():
         else:
             return out
 
+    @deprecated_function(message="Please use chords.plot_chord to have a better control of the plots")
     def plot_chords(self, all_chords=True, positive_color='blue', negative_color='green', error_color='red',
                     ax=None, lw=2, axis_labels=True):
         """ Plots the chords of the occultation
@@ -698,35 +701,9 @@ class Occultation():
             lw (int, float): linewidth of the chords. Default: 2
             axis_labels (bool): If True it prints the labels of the axis of the image.
         """
-        ax = ax or plt.gca()
-        positions = self.positions
-        for site in positions.keys():
-            pos_obs = positions[site]
-            for lc in pos_obs.keys():
-                pos_lc = pos_obs[lc]
-                if type(pos_lc) != _PositionDict:
-                    continue
-                if pos_lc['status'] == 'negative':
-                    arr = np.array([pos_lc['start_obs']['value'], pos_lc['end_obs']['value']])
-                    ax.plot(*arr.T, '--', color=negative_color, linewidth=lw)
-                else:
-                    n = 0
-                    if pos_lc['immersion']['on'] or all_chords:
-                        arr = np.array([pos_lc['immersion']['error']])
-                        ax.plot(*arr.T, color=error_color, linewidth=lw*3/2)
-                        n += 1
-                    if pos_lc['emersion']['on'] or all_chords:
-                        arr = np.array([pos_lc['emersion']['error']])
-                        ax.plot(*arr.T, color=error_color, linewidth=lw*3/2)
-                        n += 1
-                    if n == 2:
-                        arr = np.array([pos_lc['immersion']['value'], pos_lc['emersion']['value']])
-                        ax.plot(*arr.T, color=positive_color, linewidth=lw)
-        ax.invert_xaxis()
-        if axis_labels:
-            ax.set_xlabel('f (km)')
-            ax.set_ylabel('g (km)')
-        ax.axis('equal')
+        self.chords.plot_chords(segment='positive', only_able=not all_chords, color=positive_color, lw=lw, ax=ax)
+        self.chords.plot_chords(segment='error', only_able=not all_chords, color=error_color, lw=lw, ax=ax)
+        self.chords.plot_chords(segment='negative', color=negative_color, lw=lw, ax=ax, linestyle='--')
 
     def get_map_sites(self):
         """ Returns Dictionary with sites in the format required by plot_occ_map function
@@ -736,8 +713,9 @@ class Occultation():
         """
         sites = {}
         color = {'positive': 'blue', 'negative': 'red'}
-        for o, l in self.__observations:
-            sites[o.name] = [o.lon.deg, o.lat.deg, 10, 10, color[self.positions[o.name][l.name]['status']]]
+        for name, chord in self.chords.items():
+            obs = chord.observer
+            sites[name] = [obs.lon.deg, obs.lat.deg, 10, 10, color[chord.status()]]
         return sites
 
     def plot_occ_map(self, **kwargs):
@@ -853,56 +831,52 @@ class Occultation():
         The format of the files are: positions in f and g, velocities in f and g, the Julian Date of the observation,
         light curve name of the corresponding position.
         """
-        positions = self.positions
         pos = []
         neg = []
         err = []
-        for ob, lc in self.__observations:
-            status = positions[ob.name][lc.name]['status']
-            l_name = lc.name.replace(' ', '_')
+        for name, chord in self.chords.items():
+            status = chord.status()
+            l_name = name.replace(' ', '_')
             if status == 'positive':
-                im = positions[ob.name][lc.name]['immersion']
-                f, g = im['value']
-                err1, err2 = im['error']
-                f1, g1 = err1
-                f2, g2 = err2
-                vf, vg = im['vel']
-                pos.append([f, g, vf, vg, im['time'].jd, l_name+'_immersion'])
-                err.append([f1, g1, vf, vg, (im['time']-im['time_err']*u.s).jd, l_name+'_immersion_err-'])
-                err.append([f2, g2, vf, vg, (im['time']+im['time_err']*u.s).jd, l_name+'_immersion_err+'])
+                im = chord.lightcurve.immersion
+                ime = chord.lightcurve.immersion_err
+                f, g, vf, vg = chord.get_fg(time=im, vel=True)
+                f1, g1 = chord.get_fg(time=im-ime*u.s)
+                f2, g2 = chord.get_fg(time=im+ime*u.s)
+                pos.append([f, g, vf, vg, im.jd, l_name+'_immersion'])
+                err.append([f1, g1, vf, vg, (im-ime*u.s).jd, l_name+'_immersion_err-'])
+                err.append([f2, g2, vf, vg, (im+ime*u.s).jd, l_name+'_immersion_err+'])
 
-                em = positions[ob.name][lc.name]['emersion']
-                f, g = em['value']
-                err1, err2 = em['error']
-                f1, g1 = err1
-                f2, g2 = err2
-                vf, vg = em['vel']
-                pos.append([f, g, vf, vg, em['time'].jd, l_name+'_emersion'])
-                err.append([f1, g1, vf, vg, (em['time']-em['time_err']*u.s).jd, l_name+'_emersion_err-'])
-                err.append([f2, g2, vf, vg, (em['time']+em['time_err']*u.s).jd, l_name+'_emersion_err+'])
+                em = chord.lightcurve.emersion
+                eme = chord.lightcurve.emersion_err
+                f, g, vf, vg = chord.get_fg(time=em, vel=True)
+                f1, g1 = chord.get_fg(time=em-eme*u.s)
+                f2, g2 = chord.get_fg(time=em+eme*u.s)
+                pos.append([f, g, vf, vg, em.jd, l_name+'_emersion'])
+                err.append([f1, g1, vf, vg, (em-eme*u.s).jd, l_name+'_emersion_err-'])
+                err.append([f2, g2, vf, vg, (em+eme*u.s).jd, l_name+'_emersion_err+'])
+
             if status == 'negative':
-                ini = positions[ob.name][lc.name]['start_obs']
-                f, g = ini['value']
-                vf, vg = ini['vel']
-                neg.append([f, g, vf, vg, ini['time'].jd, l_name+'_start'])
+                ini = chord.lightcurve.initial_time
+                f, g, vf, vg = chord.get_fg(time=ini, vel=True)
+                neg.append([f, g, vf, vg, ini.jd, l_name+'_start'])
 
-                end = positions[ob.name][lc.name]['end_obs']
-                f, g = end['value']
-                vf, vg = end['vel']
-                neg.append([f, g, vf, vg, end['time'].jd, l_name+'_end'])
+                end = chord.lightcurve.end_time
+                f, g, vf, vg = chord.get_fg(time=end, vel=True)
+                neg.append([f, g, vf, vg, end.jd, l_name+'_end'])
         if len(pos) > 0:
             f = open('occ_{}_pos.txt'.format(self.body.shortname.replace(' ', '_')), 'w')
             for line in pos:
-                f.write('{:9.3f} {:9.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
+                f.write('{:10.3f} {:10.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
             f.close()
             f = open('occ_{}_err.txt'.format(self.body.shortname.replace(' ', '_')), 'w')
             for line in err:
-                f.write('{:9.3f} {:9.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
+                f.write('{:10.3f} {:10.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
             f.close()
         if len(neg) > 0:
             f = open('occ_{}_neg.txt'.format(self.body.shortname.replace(' ', '_')), 'w')
             for line in neg:
-                f.write('{:9.3f} {:9.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
+                f.write('{:10.3f} {:10.3f} {:-6.2f} {:-6.2f} {:16.8f} {}\n'.format(*line))
             f.close()
 
     def __str__(self):
@@ -922,20 +896,13 @@ class Occultation():
 
         count = {'positive': 0, 'negative': 0, 'visual': 0}
         string = {'positive': '', 'negative': '', 'visual': ''}
-        if len(self.__observations) > 0:
-            pos = self.positions
-            for ob, lc in self.__observations:
-                status = pos[ob.name][lc.name]['status']
-                if count[status] > 0:
-                    string[status] += '-'*79 + '\n'
-                string[status] += ob.__str__() + '\n'
-                ephem_altaz = ob.altaz(lc.time_mean, self.body.ephem.get_position(lc.time_mean))
-                string[status] += 'Target altitude: {:.1f} deg\n'.format(ephem_altaz[0])
-                string[status] += 'Target azimuth:  {:.1f} deg\n\n'.format(ephem_altaz[1])
-                string[status] += lc.__str__() + ''
+        if len(self.chords) > 0:
+            for chord in self.chords.values():
+                status = chord.status()
+                string[status] += chord.__str__() + '\n'
                 count[status] += 1
 
-        if np.sum([count[i] for i in count.keys()]) == 0:
+        if len(self.chords) == 0:
             out += 'No observations reported'
         else:
             out += '\n'.join(['{} {} observations'.format(count[k], k) for k in string.keys() if count[k] > 0])